+++ src/store/editor.ts
@@ -12,7 +12,9 @@ import {
 } from "../lib/craftNode";
 import CraftNodeResolver from "../lib/CraftNodeResolver";
 
-export type EditorState = {
+export type EditorStoreInstanceType = ReturnType<typeof useEditor>;
+
+export interface EditorState {
   nodeMap: Map<string, CraftNode>;
   rootNodes: string[];
   selectedUuid: string | null;